requirements (1).txt
openai==0.28
taskingai
streamlit
pillow